requirements.txt
keras==2.11.0
matplotlib==3.6.2
numpy==1.23.0
easydict==1.10
PyYAML==6.0